requirements.txt
# Python
attrs
ipython
twine
numba

# Scientific stack
numpy>=1.18.2
pytest>=5.4.1
scipy>=1.4.1
pandas>1.0
openpyxl>3.0
pygmo

# Notebooks
jupyter
notebook>6
nbextensions
jupytext

# Plotting
matplotlib

# Formatters and Checkers
pre-commit

# Tests
pytest-cov
pytest-xdist
pytest-regressions
codecov

# API for COVID-19
COVID19py
